LPD-X Inline KeywordQueryContributorsRegistry/KeywordQueryContributorsRegistryImpl
