--- modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/indexer/KeywordQueryContributorsRegistryImpl.java
@@ -1,56 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.search.internal.indexer;
-
-import com.liferay.osgi.service.tracker.collections.list.ServiceTrackerList;
-import com.liferay.osgi.service.tracker.collections.list.ServiceTrackerListFactory;
-import com.liferay.portal.search.spi.model.query.contributor.KeywordQueryContributor;
-
-import java.util.Collection;
-import java.util.List;
-
-import org.osgi.framework.BundleContext;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
-
-/**
- * @author André de Oliveira
- */
-@Component(service = KeywordQueryContributorsRegistry.class)
-public class KeywordQueryContributorsRegistryImpl
-	implements KeywordQueryContributorsRegistry {
-
-	@Override
-	public List<KeywordQueryContributor> filterKeywordQueryContributors(
-		Collection<String> excludes, Collection<String> includes) {
-
-		return IncludeExcludeUtil.filter(
-			_serviceTrackerList.toList(), includes, excludes,
-			this::getClassName);
-	}
-
-	@Activate
-	protected void activate(BundleContext bundleContext) {
-		_serviceTrackerList = ServiceTrackerListFactory.open(
-			bundleContext, KeywordQueryContributor.class,
-			"(!(indexer.class.name=*))");
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		_serviceTrackerList.close();
-	}
-
-	protected String getClassName(Object object) {
-		Class<?> clazz = object.getClass();
-
-		return clazz.getName();
-	}
-
-	private ServiceTrackerList<KeywordQueryContributor> _serviceTrackerList;
-
-}
