+++ modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/indexer/helper/AddSearchKeywordsQueryContributorHelperImpl.java
@@ -5,6 +5,8 @@
 
 package com.liferay.portal.search.internal.indexer.helper;
 
+import com.liferay.osgi.service.tracker.collections.list.ServiceTrackerList;
+import com.liferay.osgi.service.tracker.collections.list.ServiceTrackerListFactory;
 import com.liferay.portal.kernel.search.BooleanClauseOccur;
 import com.liferay.portal.kernel.search.BooleanQuery;
 import com.liferay.portal.kernel.search.ParseException;
@@ -13,8 +15,8 @@ import com.liferay.portal.kernel.search.generic.StringQuery;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.search.constants.SearchContextAttributes;
+import com.liferay.portal.search.internal.indexer.IncludeExcludeUtil;
 import com.liferay.portal.search.internal.indexer.IndexerProvidedClausesUtil;
-import com.liferay.portal.search.internal.indexer.KeywordQueryContributorsRegistry;
 import com.liferay.portal.search.internal.util.SearchStringUtil;
 import com.liferay.portal.search.spi.model.query.contributor.KeywordQueryContributor;
 import com.liferay.portal.search.spi.model.query.contributor.helper.KeywordQueryContributorHelper;
@@ -23,8 +25,10 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 
+import org.osgi.framework.BundleContext;
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
+import org.osgi.service.component.annotations.Deactivate;
 
 /**
  * @author André de Oliveira
@@ -44,6 +48,18 @@ public class AddSearchKeywordsQueryContributorHelperImpl
 		_addKeywordQueryContributorClauses(booleanQuery, searchContext);
 	}
 
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_serviceTrackerList = ServiceTrackerListFactory.open(
+			bundleContext, KeywordQueryContributor.class,
+			"(!(indexer.class.name=*))");
+	}
+
+	@Deactivate
+	protected void deactivate() {
+		_serviceTrackerList.close();
+	}
+
 	protected Collection<String> getStrings(
 		String string, SearchContext searchContext) {
 
@@ -51,9 +67,6 @@ public class AddSearchKeywordsQueryContributorHelperImpl
 			SearchStringUtil.splitAndUnquote(
 				(String)searchContext.getAttribute(string)));
 	}
-
-	@Reference
-	protected KeywordQueryContributorsRegistry keywordQueryContributorsRegistry;
 
 	private void _addKeywordQueryContributorClauses(
 		BooleanQuery booleanQuery, SearchContext searchContext) {
@@ -71,13 +84,15 @@ public class AddSearchKeywordsQueryContributorHelperImpl
 		}
 
 		List<KeywordQueryContributor> filteredKeywordQueryContributors =
-			keywordQueryContributorsRegistry.filterKeywordQueryContributors(
+			IncludeExcludeUtil.filter(
+				_serviceTrackerList.toList(),
+				getStrings(
+					"search.full.query.clause.contributors.includes",
+					searchContext),
 				getStrings(
 					"search.full.query.clause.contributors.excludes",
 					searchContext),
-				getStrings(
-					"search.full.query.clause.contributors.includes",
-					searchContext));
+				this::_getClassName);
 
 		for (KeywordQueryContributor keywordQueryContributor :
 				filteredKeywordQueryContributors) {
@@ -118,5 +133,13 @@ public class AddSearchKeywordsQueryContributorHelperImpl
 			throw new RuntimeException(parseException);
 		}
 	}
+
+	private String _getClassName(Object object) {
+		Class<?> clazz = object.getClass();
+
+		return clazz.getName();
+	}
+
+	private ServiceTrackerList<KeywordQueryContributor> _serviceTrackerList;
 
 }
